Synchronize dev to main

+++ AddonManagerTest/app/test_utilities.py
@@ -25,13 +25,7 @@ from datetime import datetime
 import unittest
 from unittest.mock import MagicMock, patch, mock_open
 import os
-import sys
 import subprocess
-
-try:
-    import FreeCAD
-except ImportError:
-    FreeCAD = None
 
 from AddonManagerTest.app.mocks import MockAddon as Addon
 
@@ -49,10 +43,8 @@ class TestUtilities(unittest.TestCase):
 
     @classmethod
     def tearDownClass(cls):
-        try:
+        if os.path.exists("AM_INSTALLATION_DIGEST.txt"):
             os.remove("AM_INSTALLATION_DIGEST.txt")
-        except FileNotFoundError:
-            pass
 
     def test_recognized_git_location(self):
         recognized_urls = [
